requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Django==3.1.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
gunicorn==20.0.4
idna==2.10
imutils==0.5.3
kiwisolver==1.2.0
matplotlib==3.3.1
numpy==1.19.2
opencv-python==4.4.0.42
Pillow==7.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
pyTelegramBotAPI==3.7.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
